pong/requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-friendship==1.9.6
jsonpickle==3.0.3
requests==2.31.0
pyjwt==2.8.0